Verification.
Well - good that I decided to do this verification because it kinda
ruined few plans.

It seems that the translation is not simply a files taken over
but also logic inside the codebase is slightly different as in there
is a different set of files etc.

So lets ditch the idea of pulling the original translation from the
stuff and just use Russian as a baseline. and translate from that.

Also there is an encodin stuff to be figured out.

Will test in next commit whether stuff will work if i just use utf-8 for
everything.